DCS-777This ticket has been put on hold because the overall concept of date format needs a broader rethink.
The ticket was to change the date display from short form to lolg, ag 12/08/2021 to 12 August 2021.
Although the changes made have achieved this, some failing tests suggest that some functionalities, such as saving dates to flash, may not have been suffuciently covered. Additionally there is a lack of comprehensive testing.

The pages that are impacted by the date format are
/add-court-appointment
/select-court
/request-booking
/change-time

The starategy for changing the date display was to alter the views/components/dateTimeAndCourtBriefingPicker/dateTimeAndCourtBriefingPicker.njk macro so that it could accept a moment object rather than a date as a string. The macro would then call the njk fllter  | formatDate(“D MMM YYYY) to convert the moment object to a string in the form 12 August 2021. The controller for each of the above pages would correspondingly provide a moment rather than a string.