DCS-997 Changes following PR
The prisonerSearchController and prisonerSearchValidation have been converted to typescript so that
more succinct null checking can be applied for the trim function.
Additional tests to check that whitespaces are being trimmed when search conducted.
New Nunjucks filter to trim whitespace from inputs following submission of search. Without this, any
whitesapce will be be visible in the input fields after the search has been executed.