DCS-997: changes following PR comments
New function called trimObjectValues created to remove any whitespaces around form values.

This function is used in the prisonSearchController before the call to validation. Hence the data sent to validation will not have any whitespaces.
Because of this, the tests in prisonerSearchValidation.test do not need to check for whitespaces around input values

Nunjucks trim filter no longer required either because the trim is now done via the new trim function before data sent to the ui.