DCS-997 user input trimmed to remove whitespace (#237)
* DCS-997 user input trimmed to remove whitespace

The prisonNumber, firstName and lastName have had any whitespaces trimmed. This will overcome any accidental input of whitespace for example if the user copy-pastes values from elsewhere.

* DCS-997 Changes following PR


The prisonerSearchController and prisonerSearchValidation have been converted to typescript so that
more succinct null checking can be applied for the trim function.
Additional tests to check that whitespaces are being trimmed when search conducted.
New Nunjucks filter to trim whitespace from inputs following submission of search. Without this, any
whitesapce will be be visible in the input fields after the search has been executed.

* DCS-997: changes following PR comments

New function called trimObjectValues created to remove any whitespaces around form values.

This function is used in the prisonSearchController before the call to validation. Hence the data sent to validation will not have any whitespaces.
Because of this, the tests in prisonerSearchValidation.test do not need to check for whitespaces around input values

Nunjucks trim filter no longer required either because the trim is now done via the new trim function before data sent to the ui.

* DCS-997 minor refactor

of test and nunjucks template