DCS-1090 Fixing a couple bugs (#5)
* DCS-1090 Fixing a couple bugs

When creating or amending a booking at a time that conflicts with another booking, after selecting an alternative and clicking back, the user was previously presented a not available page. Now redirecting back to the appropriate 'enter details' page (as the booking is actually present)

Also fixing issue where we were entering the edit booking flow by the wrong entry point. Needed to go via 'start-editing-booking so any previous cookie state would be cleared out. This was instead potentially showing the wrong booking

* DCS-1090 fix correct link